refactor: Repackaging player related logic.
