+++ adventures.and.beyond/src/entity/player/PlayerConstant.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.player;
 
 public class PlayerConstant {
