+++ adventures.and.beyond/src/entity/player/PlayerSpriteManager.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.player;
 
 import directionEnum.Direction;
