+++ adventures.and.beyond/src/object/interactable/objects/InteractableObject.java
@@ -1,6 +1,6 @@
 package object.interactable.objects;
 
-import entity.Player;
+import entity.player.Player;
 import helper.DrawHelper;
 import helper.Loader;
