+++ adventures.and.beyond/src/tile/TileManager.java
@@ -1,6 +1,6 @@
 package tile;
 
-import entity.Player;
+import entity.player.Player;
 import helper.Constant;
 import helper.Loader;
 import main.GamePanel;
