+++ adventures.and.beyond/src/ui/notifier/PlayerDetailNotifier.java
@@ -1,6 +1,6 @@
 package ui.notifier;
 
-import entity.Player;
+import entity.player.Player;
 import main.GamePanel;
 
 import java.awt.*;
